Create Openid_Config_Check.py
Especially during mobile testing, we may occasionally encounter the "OpenID Connect (OIDC)" protocol. In these processes, analyzing the "openid-configuration" file associated with the OpenID Connect (OIDC) protocol is crucial to understand the architecture. It is extremely important to analyze how secure the configurations in this file are. For this reason, I have developed a very simple Python script that analyzes the "openid-configuration" file. I hope it will be helpful for you.